Website Redesign after CSS Lessons
The website underwent a complete redesign after the latest lessons of
CSS I have gone through on The Odin Project. After learning the box
method, inline vs block and how they all go together, I made some
changes. (It was also a part of the assignment)
All html files / pages had a visual overhaul as I implemented a color
pallette and added it into the styles.css file. Created a navBar,
restyled the website "logo", added a footer, and centered the images on
the recipe pages.

NavBar buttons have a change of color on hover, colors may be adjusted
later down the road but it looks good for now.

This is all an occumulation of The Odin Project's lessons but also some
CSS tips and tricks I picked up from some youtube videos.

- AP